Add `nexus_capabilities::FirewallRules`
